feat: index assignment, register widening u8→u16, O(1) append/sort/reverse via Arc COW
- Parser: list[i] = val and list[i] += val as assignment targets
- Typechecker: mutability enforcement for index/field assignment targets
- Register allocator: internal u16 types with graceful error returns
- LirCell.registers widened from u8 to u16
- VM check_register widened from u8 to u16
- Value derives Default (Null variant) enabling std::mem::take
- Intrinsic append/sort/reverse use take + Arc::make_mut for COW
- 5,333 tests passing, zero clippy warnings

+++ rust/lumen-compiler/src/compiler/lir.rs
@@ -504,7 +504,7 @@ pub struct LirCell {
     pub name: String,
     pub params: Vec<LirParam>,
     pub returns: Option<String>,
-    pub registers: u8,
+    pub registers: u16,
     pub constants: Vec<Constant>,
     pub instructions: Vec<Instruction>,
     /// Metadata for effect handler scopes pushed by HandlePush instructions in this cell.
